refactor(entity): split cell_entity/mod.rs (845 lines) along natural seams
Extract cohesive subsystems into siblings:

- bandolier.rs: BandolierItem + bandolier helpers (75 lines)
- state_flags.rs: BSF constants + bitflag manipulation (74 lines)
- mod.rs: CellEntity struct + core impl (696 lines, under soft cap)
